Check if `is_active` column exists before dropping

+++ database/migrations/2021_06_01_153459_add_is_active_to_governments_table.php
@@ -14,7 +14,7 @@ class AddIsActiveToGovernmentsTable extends Migration
     public function up()
     {
         Schema::table('governments', function (Blueprint $table) {
-                $table->boolean('is_active')->default(false);
+            $table->boolean('is_active')->default(false);
         });
     }
 
@@ -25,8 +25,10 @@ class AddIsActiveToGovernmentsTable extends Migration
      */
     public function down()
     {
-        Schema::table('governments', function (Blueprint $table) {
+        if (Schema::hasColumn('governments', 'is_active')) {
+            Schema::table('governments', function (Blueprint $table) {
                 $table->dropColumn('is_active');
-        });
+            });
+        }
     }
 }
